--- src/main/java/usecases/dataStorage/WeightSchemeDsModel.java
@@ -1,3 +1,0 @@
-package usecases.dataStorage;
-
-public class WeightSchemeDsModel { }
