+++ src/main/java/usecases/dataStorage/weightScheme/OrderedWeightDsModel.java
@@ -1,4 +1,4 @@
-package usecases.dataStorage;
+package usecases.dataStorage.weightScheme;
 
 public class OrderedWeightDsModel extends WeightSchemeDsModel{
     private WeightDsModel[] orderedWeightModels;
