+++ src/main/java/usecases/dataStorage/weightScheme/WeightSchemeDsModel.java
@@ -1,0 +1,3 @@
+package usecases.dataStorage.weightScheme;
+
+public class WeightSchemeDsModel { }
